requirements.txt
asgiref==3.5.2
build==0.10.0
certifi==2022.9.24
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
gunicorn==20.1.0
packaging==23.0
Pillow==9.2.0
pyproject_hooks==1.0.0
pytz==2022.5
sqlparse==0.4.3